python/pyproject.toml
[tool.uv.workspace]
members = ["packages/autogen-kafka-extension","packages/exemple"]

[tool.uv.sources]
autogen-kafka-extension = { workspace = true }
exemple = { workspace = true }

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "autogen-kafka"
version = "0.1.0"
description = "autogen kafka support extension"
keywords = ["autogen", "kafka", "extension"]
readme = "README.md"
license = {file="LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "autogen>=0.6.1",
    "autogen-core>=0.6.1",
    "kstreams>=0.26.9",
    "confluent-kafka>=2.10.1",
    "confluent-kafka[json]>=2.10.1",
    "azure-core>=1.34.0",
    "autogen-ext[grpc]>=0.6.1",
    "aiorun>=2017.10.2",
    "grpcio>=1.70.0",
    "aiorun>=2025.1.1",
    "hatchling>=1.27.0",
    "kstreams>=0.26.9",
    "pyautogen>=0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "testcontainers>=4.10.0",
    "autogen-ext>=0.6.1",
]
